feat(main): add settings window
