+++ src/main/app_windows/index.ts
@@ -1,2 +1,3 @@
 export { CreateAboutWindow } from './about_window';
 export { CreateKeyPinWindow } from './key_pin_window';
+export { CreateSettingsWindow } from './settings_window';
